micro/book_service/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
mysqlclient==2.2.6
django-cors-headers==4.6.0
